requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.2
django-social-share==2.3.0
django-taggit==6.1.0
django_paystack==1.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
pillow==11.0.0
pypaystack2==2.1.0
python-decouple==3.8
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.1
